+++ src/http/gui_backend/private_rpc_backend.erl
@@ -37,18 +37,38 @@ handle(<<"getUserAlias">>, _) ->
     {ok, #onedata_user{
         alias = Alias
     }} = user_logic:get_user(UserId),
-    case str_utils:to_binary(Alias) of
-        <<"">> ->
-            {ok, null};
-        Bin ->
-            {ok, Bin}
+    UserAlias = case str_utils:to_binary(Alias) of
+        <<"">> -> null;
+        Bin -> Bin
+    end,
+    {ok, [
+        {<<"userAlias">>, UserAlias}
+    ]};
+
+handle(<<"changePassword">>, Props) ->
+    UserId = g_session:get_user_id(),
+    {ok, #onedata_user{
+        login = Login
+    }} = user_logic:get_user(UserId),
+    OldPassword = proplists:get_value(<<"oldPassword">>, Props),
+    NewPassword = proplists:get_value(<<"newPassword">>, Props),
+    case user_logic:change_user_password(Login, OldPassword, NewPassword) of
+        ok ->
+            ok;
+        {error, Binary} when is_binary(Binary) ->
+            gui_error:report_warning(Binary);
+        _ ->
+            gui_error:report_warning(
+                <<"Cannot change user password - old password incorrect.">>)
     end;
 
 handle(<<"setUserAlias">>, [{<<"userAlias">>, NewAlias}]) ->
     UserId = g_session:get_user_id(),
     case user_logic:modify(UserId, [{alias, NewAlias}]) of
         ok ->
-            {ok, NewAlias};
+            {ok, [
+                {<<"userAlias">>, NewAlias}
+            ]};
         {error, disallowed_prefix} ->
             gui_error:report_warning(
                 <<"Alias cannot start with \"", ?NO_ALIAS_UUID_PREFIX, "\".">>);
@@ -73,15 +93,22 @@ handle(<<"getConnectAccountEndpoint">>, [{<<"provider">>, ProviderBin}]) ->
     Provider = binary_to_atom(ProviderBin, utf8),
     HandlerModule = auth_config:get_provider_module(Provider),
     {ok, URL} = HandlerModule:get_redirect_url(true),
-    {ok, URL};
+    {ok, [
+        {<<"url">>, URL}
+    ]};
 
-handle(<<"getSupportToken">>, [{<<"spaceId">>, SpaceId}]) ->
+handle(<<"getTokenProviderSupportSpace">>, [{<<"spaceId">>, SpaceId}]) ->
     Client = #client{type = user, id = g_session:get_user_id()},
     {ok, Token} = token_logic:create(
         Client, space_support_token, {space, SpaceId}),
-    {ok, Token};
+    {ok, [
+        {<<"token">>, Token}
+    ]};
 
-handle(<<"getRedirectURL">>, [{<<"providerId">>, ProviderId}]) ->
+handle(<<"getProviderRedirectURL">>, [{<<"providerId">>, ProviderId}]) ->
     UserId = g_session:get_user_id(),
     % @todo check if provider is online, if not push update of model
-    auth_logic:get_redirection_uri(UserId, ProviderId).
+    {ok, URL} = auth_logic:get_redirection_uri(UserId, ProviderId),
+    {ok, [
+        {<<"url">>, URL}
+    ]}.