+++ src/logic/group_logic.erl
@@ -17,7 +17,7 @@
 %% API
 -export([exists/1, has_user/2, has_effective_user/2, has_effective_privilege/3,
     has_nested_group/2]).
--export([create/3, modify/2, join/2, set_privileges/3, join_group/2]).
+-export([create/3, modify/2, add_user/2, join/2, set_privileges/3, join_group/2]).
 -export([get_data/1, get_users/1, get_effective_users/1, get_spaces/1, get_providers/1,
     get_user/2, get_privileges/2, get_effective_privileges/2, get_nested_groups/1,
     get_nested_group/2, get_nested_group_privileges/2, set_nested_group_privileges/3,
@@ -162,17 +162,17 @@ modify(GroupId, Data) ->
     ok.
 
 %%--------------------------------------------------------------------
-%% @doc Adds user to a group identified by a token.
-%% Throws exception when call to the datastore fails, or token/user/group_from_token
-%% doesn't exist in db.
+%% @doc Adds user to a group. Does not check authorization - use join/2 for
+%% user adding based on authorization!
+%% Throws exception when call to the datastore fails.
 %% @end
 %%--------------------------------------------------------------------
--spec join(UserId :: binary(), Macaroon :: macaroon:macaroon()) ->
-    {ok, GroupId :: binary()}.
-join(UserId, Macaroon) ->
-    {ok, {group, GroupId}} = token_logic:consume(Macaroon),
+-spec add_user(GroupId :: binary(), UserId :: binary()) ->
+    ok.
+add_user(GroupId, UserId) ->
     case has_user(GroupId, UserId) of
-        true -> ok;
+        true ->
+            ok;
         false ->
             Privileges = privileges:group_user(),
             {ok, _} = user_group:update(GroupId, fun(Group) ->
@@ -182,8 +182,21 @@ join(UserId, Macaroon) ->
             {ok, _} = onedata_user:update(UserId, fun(User) ->
                 #onedata_user{groups = Groups} = User,
                 {ok, User#onedata_user{groups = [GroupId | Groups]}}
-            end)
-    end,
+            end),
+            ok
+    end.
+
+%%--------------------------------------------------------------------
+%% @doc Adds user to a group identified by a token.
+%% Throws exception when call to the datastore fails, or token/user/group_from_token
+%% doesn't exist in db.
+%% @end
+%%--------------------------------------------------------------------
+-spec join(UserId :: binary(), Macaroon :: macaroon:macaroon()) ->
+    {ok, GroupId :: binary()}.
+join(UserId, Macaroon) ->
+    {ok, {group, GroupId}} = token_logic:consume(Macaroon),
+    ok = add_user(GroupId, UserId),
     {ok, GroupId}.
 
 %%--------------------------------------------------------------------