+++ src/logic/token_logic.erl
@@ -19,8 +19,7 @@
 
 %% Atoms representing types of valid tokens.
 -type token_type() :: group_invite_token | group_invite_group_token | space_create_token |
-space_invite_user_token | space_invite_group_token | accounts_merge_token |
-space_support_token.
+space_invite_user_token | space_invite_group_token | space_support_token.
 
 %% Atoms representing valid resource types.
 -type resource_type() :: user | group | space.